Simplify chunk (#132)
* gpu/chunk_impl: add test for reshape_map when local data dim > 1

to test the reshape_map when local data shape is 2D

* Simplify reshape_map to avoid unnecessary rem ops.

When computing the last lid or tid, we do not need to rem operation.

Adding rem ops will cause extra performance overhead.

For example, we notice that there are around 20% overhead in
layernorm_forward kernel in llm_rs example.

simplify reshape_map to avoid rem for local id

* Define MapContinuousLinear to replace MapLinear

* examples/llm-rs: Avoid using old MapLinear which may cause unnecessary cost due to rem op

* fix clippy

+++ examples/reduce/src/lib.rs
@@ -63,7 +63,7 @@ mod tests {
                 .new_tensor_view(partial_sums.as_mut_slice())
                 .expect("alloc failed");
             let smem = block_dim * core::mem::size_of::<u32>() as u32;
-            let config = gpu_host::gpu_config!(grid_dim, 1, 1, block_dim, 1, 1, smem as u32);
+            let config = gpu_host::gpu_config!(grid_dim, 1, 1, block_dim, 1, 1, smem);
             reduce_per_grid::launch(config, ctx, m, &d_inputs, &mut d_partial_sums)
                 .expect("reduce_per_grid kernel launch failed");
             d_partial_sums
